added BaseController.cs
